requirements.txt
# crewai
# crewai-tools
# google-generativeai
# python-dotenv
# pydantic
# pathlib
# typing-extensions

crewai>=0.28.0
crewai-tools>=0.1.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pathlib>=1.0.1
typing-extensions>=4.0.0
langchain-google-genai>=0.0.2